requirements.txt
filelock==3.18.0
fsspec==2025.3.2
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
sympy==1.14.0
torch==2.7.0
typing_extensions==4.13.2
filelock==3.18.0
fsspec==2025.3.2
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
PyYAML==6.0.2
sympy==1.14.0
torch==2.7.0
typing_extensions==4.13.2
